* client-src/client_util.c: Fix compiler warning * config/amanda/flags.m4: Fix egrep expresion for solaris
git-svn-id: https://svn.code.sf.net/p/amanda/code/amanda/trunk@6356 a8d146d6-cc15-0410-8900-af154a0219e0

+++ client-src/client_util.c
@@ -1677,11 +1677,11 @@ check_access(
 
     if(EUIDACCESS(filename, mode) == -1) {
 	g_printf(_("ERROR [can not %s %s: %s (ruid:%d euid:%d)\n"), noun, quoted, strerror(errno),
-	    getuid(), geteuid());
+	    (int)getuid(), (int)geteuid());
 	result = FALSE;
     } else {
 	g_printf(_("OK %s %s (ruid:%d euid:%d)\n"), quoted, adjective,
-	    getuid(), geteuid());
+	    (int)getuid(), (int)geteuid());
 	result = TRUE;
     }
     amfree(quoted);
